payment controller completed

+++ Models/Staff.cs
@@ -12,7 +12,7 @@ namespace Thakshilawa.Models
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public float StaffId { get; set; }
+        public int StaffId { get; set; }
         public string Name { get; set; }
 
         public string NICNo { get; set; }
